Put created in userController

+++ app/Controllers/Http/UserController.js
@@ -5,7 +5,6 @@ const User = use("App/Models/User");
 class UserController {
   async create({ request, auth, response }) {
     const data = request.only(["name", "email", "password", "type"]);
-    console.log('type user->', auth.user.type)
     if (auth.user.type !== "admin") {
       return response
         .status(401)
@@ -38,12 +37,27 @@ class UserController {
   async destroy({ params, auth, response }) {
     const user = await User.find(params.id);
 
-    // if (auth.user.type !== "admin") {
-    //   return response
-    //     .status(401)
-    //     .send({ error: "Usuario não aturizado para essa operação" });
-    // } 
+    if (auth.user.type !== "admin") {
+      return response
+        .status(401)
+        .send({ error: "Usuario não aturizado para essa operação" });
+    } 
+    
     return user.delete();
+  }
+  async update({ params, auth, response, request}) {
+    const user = await User.find(params.id);
+
+    const data  = request.only(["password", "name", "email", "type"])
+
+    if (auth.user.type !== "admin") {
+      return response
+        .status(401)
+        .send({ error: "Usuario não aturizado para essa operação" });
+    } 
+    user.merge(data);
+    await user.save(data);
+    return user;
   }
 }
